Add new benchmarking charts and other cleanups
