+++ v3data/constants.py
@@ -1,0 +1,1 @@
+WETH_ADDRESS = "0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2"
